+++ multiverse/search.php
@@ -36,8 +36,7 @@ if ($pag_tot > 1) {
     <div id="main">
       <article id="page">
         <h1><?php if (isArchive()) {
-          $date = $_zp_current_search->getSearchDate();
-          echo '<a href="' . getCustomPageURL('archive') . '">' . gettext('Gallery archive') . '</a>' . ': ' . getFormattedLocaleDate('F Y', $date);
+          echo '<a href="' . getCustomPageURL('archive') . '">' . gettext('Gallery archive') . '</a>' . ': ' . getSearchDate();
         } else {
           echo gettext_th('Search results');
         }